+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log:
 
+## 1.3.1
+
+    - Se ha añadido una extensión para comprobar si un struct es 'null'
+
+## 1.3.0
+    
+    - Se ha modificado 'Singleton.cs' para que no cree un objeto si no se ha encontrado, en vez de eso lanza un 'Exception'
+
+## 1.2.2
+
+    - Se han añadido dos extensiones, 'Contains' y 'RecursiveFindChild'
+
+## 1.2.1
+
+    - Se ha actualizado 'SafeArea.cs'
+
 ## 1.2.0
 
     - Se ha añadido NonDrawingGraphic, una imagen que no dibuja ningún sprite, solo capta raycasts. Perfecto para botones.